requirements.txt
black==20.8b1
isort==5.7.0
pandas==1.2.1
pandas-gbq==0.14.1 
plotly==4.14.3
streamlit==0.75.0
google-oauth2-tool==0.0.3
psycopg2-binary
python-dotenv